+++ app/src/main/java/com/example/cinema_shift_2024/CinemaApplication.kt
@@ -2,9 +2,9 @@ package com.example.cinema_shift_2024
 
 import android.app.Application
 import com.example.cinema_shift_2024.details.di.detailsModule
+import com.example.cinema_shift_2024.feature.posters.di.postersModule
 import com.example.cinema_shift_2024.main.di.mainModule
 import com.example.cinema_shift_2024.navigation.navigationModule
-import com.example.cinema_shift_2024.posters.di.postersModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
