+++ app/src/main/java/com/example/cinema_shift_2024/main/presentation/NavigationOption.kt
@@ -1,6 +1,6 @@
 package com.example.cinema_shift_2024.main.presentation
 
-import com.example.cinema_shift_2024.posters.PostersRoute
+import com.example.cinema_shift_2024.feature.posters.PostersRoute
 import com.example.cinema_shift_2024.tickets.TicketsRoute
 import kotlin.reflect.KClass
