+++ app/src/main/java/com/example/cinema_shift_2024/main/ui/MainScreen.kt
@@ -27,10 +27,10 @@ import androidx.navigation.toRoute
 import com.example.cinema_shift_2024.R
 import com.example.cinema_shift_2024.details.DetailsRoute
 import com.example.cinema_shift_2024.details.ui.DetailsScreen
+import com.example.cinema_shift_2024.feature.posters.PostersRoute
 import com.example.cinema_shift_2024.main.presentation.MainViewModel
 import com.example.cinema_shift_2024.main.presentation.NavigationOption
 import com.example.cinema_shift_2024.navigation.NavControllerHolder
-import com.example.cinema_shift_2024.posters.PostersRoute
 import com.example.cinema_shift_2024.posters.ui.PostersScreen
 import com.example.cinema_shift_2024.tickets.TicketsRoute
 import com.example.cinema_shift_2024.tickets.ui.TicketsScreen
