+++ app/src/main/java/com/example/cinema_shift_2024/navigation/NavigationModule.kt
@@ -1,11 +1,11 @@
 package com.example.cinema_shift_2024.navigation
 
 import com.example.cinema_shift_2024.details.presentation.DetailsRouter
+import com.example.cinema_shift_2024.feature.posters.presentation.PostersRouter
 import com.example.cinema_shift_2024.main.presentation.MainRouter
 import com.example.cinema_shift_2024.navigation.featurerouter.DetailsRouterImpl
 import com.example.cinema_shift_2024.navigation.featurerouter.MainRouterImpl
 import com.example.cinema_shift_2024.navigation.featurerouter.PostersRouterImpl
-import com.example.cinema_shift_2024.posters.presentation.PostersRouter
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.binds
