+++ app/src/main/java/com/example/cinema_shift_2024/navigation/featurerouter/MainRouterImpl.kt
@@ -1,8 +1,8 @@
 package com.example.cinema_shift_2024.navigation.featurerouter
 
+import com.example.cinema_shift_2024.feature.posters.PostersRoute
 import com.example.cinema_shift_2024.main.presentation.MainRouter
 import com.example.cinema_shift_2024.navigation.GlobalRouter
-import com.example.cinema_shift_2024.posters.PostersRoute
 import com.example.cinema_shift_2024.tickets.TicketsRoute
 
 class MainRouterImpl(
