+++ app/src/main/java/com/example/cinema_shift_2024/navigation/featurerouter/PostersRouterImpl.kt
@@ -1,8 +1,8 @@
 package com.example.cinema_shift_2024.navigation.featurerouter
 
 import com.example.cinema_shift_2024.details.DetailsRoute
+import com.example.cinema_shift_2024.feature.posters.presentation.PostersRouter
 import com.example.cinema_shift_2024.navigation.GlobalRouter
-import com.example.cinema_shift_2024.posters.presentation.PostersRouter
 
 class PostersRouterImpl(
     private val router: GlobalRouter
