--- app/src/main/java/com/example/cinema_shift_2024/posters/PostersRoute.kt
@@ -1,6 +1,0 @@
-package com.example.cinema_shift_2024.posters
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object PostersRoute
