--- app/src/main/java/com/example/cinema_shift_2024/posters/data/network/PostersApi.kt
@@ -1,10 +1,0 @@
-package com.example.cinema_shift_2024.posters.data.network
-
-import com.example.cinema_shift_2024.posters.data.model.FilmsResponse
-import retrofit2.http.GET
-
-interface PostersApi {
-
-    @GET("/cinema/today")
-    suspend fun getFilmsList(): FilmsResponse
-}
