--- app/src/main/java/com/example/cinema_shift_2024/posters/domain/repository/PostersRepository.kt
@@ -1,8 +1,0 @@
-package com.example.cinema_shift_2024.posters.domain.repository
-
-import com.example.cinema_shift_2024.posters.domain.entity.Film
-
-interface PostersRepository {
-
-    suspend fun getList(): List<Film>
-}
