--- app/src/main/java/com/example/cinema_shift_2024/posters/domain/usecase/GetListUseCase.kt
@@ -1,9 +1,0 @@
-package com.example.cinema_shift_2024.posters.domain.usecase
-
-import com.example.cinema_shift_2024.posters.domain.entity.Film
-import com.example.cinema_shift_2024.posters.domain.repository.PostersRepository
-
-class GetListUseCase(private val repository: PostersRepository) {
-
-    suspend operator fun invoke(): List<Film> = repository.getList()
-}
