--- app/src/main/java/com/example/cinema_shift_2024/posters/presentation/PostersRouter.kt
@@ -1,6 +1,0 @@
-package com.example.cinema_shift_2024.posters.presentation
-
-interface PostersRouter {
-
-    fun openDetails(filmId: String)
-}
