+++ app/src/main/res/values/strings.xml
@@ -2,11 +2,7 @@
     <string name="app_name">cinema-shift-2024</string>
 
 
-    <string name="posters_title">Афиша на сегодня</string>
-    <string name="film_title_runtime_age">%1$s - %2$s минут. %3$s</string>
-    <string name="kinopoisk_rating">Оценка на Кинопоиск: %1$s</string>
-    <string name="imdb_rating">Оценка на IMDB: %1$s</string>
-    <string name="show_details">Подробнее</string>
+
 
     <string name="realise_date">Дата выхода - %1$s</string>
     <string name="actors">Актёры: %1$s</string>
