+++ feature/posters/src/main/kotlin/com/example/cinema_shift_2024/feature/posters/PostersRoute.kt
@@ -1,0 +1,6 @@
+package com.example.cinema_shift_2024.feature.posters
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object PostersRoute
