+++ feature/posters/src/main/kotlin/com/example/cinema_shift_2024/feature/posters/presentation/PostersRouter.kt
@@ -1,0 +1,6 @@
+package com.example.cinema_shift_2024.feature.posters.presentation
+
+interface PostersRouter {
+
+    fun openDetails(filmId: String)
+}
