+++ feature/posters/src/main/kotlin/com/example/cinema_shift_2024/feature/posters/ui/LoadingComponent.kt
